+++ src/index.ts
@@ -3,6 +3,7 @@ import { Button } from './components/Button'
 import { ButtonGroup } from './components/ButtonGroup'
 import { ContactLink } from './components/ContactLink'
 import { Modal } from './components/Modal'
+import { Placeholder } from './components/Placeholder'
 import { Spinner } from './components/Spinner'
 import { TMode, TShape } from './typings/types'
 
@@ -12,9 +13,9 @@ export {
   ButtonGroup,
   ContactLink,
   Modal,
+  Placeholder,
   Spinner
 }
-
 export type {
   TMode,
   TShape